Fixed bug in game 1 + enabled taking damage

+++ DIY Boss Rush Game/Game1.cs
@@ -331,11 +331,11 @@ namespace DIY_Boss_Rush_Game
                 _spriteBatch.Draw(uiBossMain, new Vector2(0, 0), Color.White);
 
                 //REPLACEMENTS: boss and player current/max health properties || Count: 5 unresolved
-                _spriteBatch.Draw(uiPlayerBar, new Vector2(0), new Rectangle(0, 0, (int)(143 + 285 * (float)(player.CurrHealth/player.MaxHealth)), 1080), Color.White);
-                _spriteBatch.Draw(uiPlayerNub, new Vector2(0), new Rectangle(0 + (int)(285 - 285 * (float)(player.CurrHealth / player.MaxHealth)),0,1920,1080), Color.White);
-                _spriteBatch.Draw(uiBossBar, new Vector2(1492 + (285 - (285 * (float)(boss[0].CurrHealth / boss[0].MaxHealth))), 0),
-                    new Rectangle((int)(1492 + (285 - (285 * (float)(boss[0].CurrHealth / boss[0].MaxHealth)))),0,428,1080), Color.White);
-                _spriteBatch.Draw(uiBossNub, new Vector2(1479 + (285 - (285 * (float)(boss[0].CurrHealth / boss[0].MaxHealth))), -1), 
+                _spriteBatch.Draw(uiPlayerBar, new Vector2(0), new Rectangle(0, 0, (int)(143 + 285 * (float)player.CurrHealth/(float)player.MaxHealth), 1080), Color.White);
+                _spriteBatch.Draw(uiPlayerNub, new Vector2(0), new Rectangle(0 + (int)(285 - 285 * (float)player.CurrHealth / (float)player.MaxHealth), 0,1920,1080), Color.White);
+                _spriteBatch.Draw(uiBossBar, new Vector2(1492 + (285 - (285 * (float)boss[0].CurrHealth / (float)boss[0].MaxHealth)), 0),
+                    new Rectangle((int)(1492 + (285 - (285 * (float)boss[0].CurrHealth / (float)boss[0].MaxHealth))),0,428,1080), Color.White);
+                _spriteBatch.Draw(uiBossNub, new Vector2(1479 + (285 - (285 * (float)boss[0].CurrHealth / (float)boss[0].MaxHealth)), -1), 
                     new Rectangle(1478, 0,14,1080), Color.White);
                 _spriteBatch.Draw(uiPlayerTop, new Vector2(0, 0), Color.White);
                 _spriteBatch.Draw(uiBossTop, new Vector2(0, 0), Color.White);
